backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9
requests==2.32.3
pillow==10.0.0
transformers==4.42.2
torch